--- MLS/MLS/Extensions/UIViewController+.swift
@@ -1,14 +1,0 @@
-//
-//  UIViewController+.swift
-//  MLS
-//
-//  Created by SeoJunYoung on 2/21/24.
-//
-
-import UIKit
-
-extension UIViewController {
-    func checkPopUpView(completion: @escaping () -> Void) {
-        completion()
-    }
-}
