+++ MLS/MLS/Helper/Controllers/BasicController.swift
@@ -10,7 +10,7 @@ import UIKit
 import SnapKit
 
 class BasicController: UIViewController {
-   
+    
     init() {
         super.init(nibName: nil, bundle: nil)
         print(self, "init")
@@ -34,47 +34,24 @@ extension BasicController {
 
 // MARK: - SetUp
 private extension BasicController {
-
+    
     func setUpColor() {
         view.backgroundColor = .systemBackground
     }
 }
 
 // MARK: - Methods
-
 extension BasicController {
-    
-    func showLaunchScreen(completion: @escaping () -> Void) {
-        let launchImageView: UIImageView = {
-            let view = UIImageView()
-            guard let gifURL = Bundle.main.url(forResource: "launchScreen", withExtension: "gif"),
-                  let gifData = try? Data(contentsOf: gifURL),
-                  let source = CGImageSourceCreateWithData(gifData as CFData, nil)
-            else { return view }
-            let frameCount = CGImageSourceGetCount(source)
-            var images = [UIImage]()
-
-            (0..<frameCount)
-                .compactMap { CGImageSourceCreateImageAtIndex(source, $0, nil) }
-                .forEach { images.append(UIImage(cgImage: $0)) }
-
-            view.animationImages = images
-            view.animationDuration = 2
-            view.animationRepeatCount = 1
-            return view
-        }()
-        
-        view.addSubview(launchImageView)
-        launchImageView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        launchImageView.startAnimating()
-    
-        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
-            launchImageView.removeFromSuperview()
-            launchImageView.stopAnimating()
-            completion()
+    func checkAnnouncementData() {
+        let manager = UserDefaultsManager()
+        let notice = AnnouncementData(
+            number: 2,
+            title: "공지사항",
+            content: "안녕하세요 메랜사 입니다."
+        )
+        if !manager.fetchIsCheckNotice(number: notice.number) {
+            PopUpMaker.showPopUp(title: notice.title, content: notice.content)
+            manager.setIsCheckNotice(toggle: true, number: notice.number)
         }
     }
 }
-
