+++ MLS/MLS/Resource/SceneDelegate.swift
@@ -15,7 +15,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         
         window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = UINavigationController(rootViewController: DictLandingViewController(viewModel: DictLandingViewModel()))
+        let rootVC = DictLandingViewController(viewModel: DictLandingViewModel())
+        window?.rootViewController = UINavigationController(rootViewController: LaunchViewController(rootVC: rootVC))
         window?.makeKeyAndVisible()
     }
     
@@ -37,19 +38,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         if checkUpdateAvailable() {
             presentUpdateAlertVC(rootVC: rootVC)
         }
-        let manager = UserDefaultsManager()
-        let notice = AnnouncementData(
-            number: 0,
-            title: "공지사항",
-            content: "안녕하세요 메랜사 입니다."
-        )
-        rootVC.checkPopUpView {
-            if !manager.fetchIsCheckNotice(number: notice.number) {
-                PopUpMaker.showPopUp(title: notice.title, content: notice.content)
-                manager.setIsCheckNotice(toggle: true, number: notice.number)
-            }
-        }
-        
     }
     
     func sceneDidEnterBackground(_ scene: UIScene) {}
