+++ MLS/MLS/Scenes/DictionaryPageScene/DictLanding/DictLandingViewController.swift
@@ -59,17 +59,12 @@ class DictLandingViewController: BasicController {
 extension DictLandingViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
-        print(#function, "start")
-        showLaunchScreen { [weak self] in
-            guard let self = self else { return }
-            self.bind()
-            self.setUp()
-        }
-        print(#function, "stop")
+        bind()
+        setUp()
+        checkAnnouncementData()
     }
     
     override func viewWillAppear(_ animated: Bool) {
-        print(#function, "start")
         self.navigationController?.navigationBar.isHidden = true
         IndicatorManager.showIndicator(vc: self)
         viewModel.fetchSectionDatas() {
